Refactor Youtube video
