+++ src/Components/VideoPlayer/VideoPlayer.js
@@ -6,7 +6,7 @@ const styles = {
   buttonContainer: 'w-full absolute text-white z-40 self-end flex justify-end',
 
   buttonSkipVideo:
-    'bg-gradient-to-r from-red-400 to-yellow-500 font-semibold rounded-full w-24 h-10 text-sm self-end uppercase mr-5 md:mr-10 mb-5 md:mb-9 hover:from-red-500 hover:to-yellow-600'
+    'bg-gradient-to-r from-red-600 via-red-500 to-red-400 font-semibold rounded-full w-24 h-10 text-sm self-end uppercase mr-5 md:mr-10 mb-5 md:mb-9 hover:from-red-700 hover:via-red-600 hover:to-red-500'
 }
 
 export default function VideoPlayer({ url, playing }) {
@@ -18,6 +18,13 @@ export default function VideoPlayer({ url, playing }) {
         playing={playing}
         width="100%"
         height="100vh"
+        config={{
+          youtube: {
+            controls: 0,
+            showinfo: 0,
+            autohide: 1
+          }
+        }}
       />
 
       <div className={styles.buttonContainer}>
